PHPCS: rename ruleset to use a standard PHPCS ruleset name
If a PHPCS ruleset is called `.phpcs.xml` `phpcs.xml`, `.phpcs.xml.dist` or `phpcs.xml.dist`, PHPCS will automatically recognize it and use it.

PHPCS uses the loading order as per the order in which I've listed the file names above.

Using `phpcs.xml.dist` as a filename will therefore allow for individual developers to have a custom ruleset with one of those other names which then automatically overrules the project ruleset (though the dev custom ruleset can include the project ruleset as a basis, which is the recommended practice).

Includes adding the filenames which can be used to overrule the ruleset to the `.gitignore` file.