Travis: use Composer to install PHPCS
As it was, the Travis script installed a very specific version of PHP_CodeSniffer and WordPressCS, without any check that these are in line with each other dependency-wise.

Using Composer to install PHPCS and WPCS (as well as PHPUnit), will ensure that the versions installed are always in line with each other and will solve the current build failure.

Includes removing the custom directories which would be created from the `.gitignore` file.
Includes minor tweaking of the directories being cached for Composer by Travis.