Travis: make the script more efficient
In the previous commit, I added a Composer command to install PHPCS.

However, this will also install PHPUnit, which was already being installed globally a few lines earlier.

Secondly, PHPCS will give the same result across PHP versions, so running it on all four builds is highly inefficient. You only really need to run it once.

This commit adds an environment variable `SNIFF` to toggle the running of PHPCS on for just one build in the script.

Along those same lines, the full `composer install` only needs to be run when PHPCS is needed, for the other runs, just installing PHPUnit globally is sufficient.