Composer: update Composer PHPCS plugin dependency
A new version has been released of the DealerDirect PHPCS Composer installer.
As minor releases pre-1.0 are treated as major releases, the version needs to be bumped for the newer version to be installed.

Note: the `composer update` command has been run with `--with-all-dependencies`, which means that PHP_CodeSniffer has also been updated to the latest released version.

Ref:
* https://github.com/Dealerdirect/phpcodesniffer-composer-installer/releases/tag/v0.5.0
* https://github.com/squizlabs/PHP_CodeSniffer/releases/tag/3.4.0