Travis/PHPCS: consolidate the PHPCS runs
The Travis script previously contained two PHPCS runs, one for the `plugins` directory, one for the `themes` directory.

The ruleset file already contained a `<file>` directive, but this was set to `.` (= all files).

By making this `<file>` directive more specific, the `phpcs` command can be run from the project root and will still only examine the `plugins` and `themes` directories.

**_Note_**: As a suggestion for the future: you may want to add the `tests` directory as well.