PHPCS: add checking for PHP cross-version compatibility
This adds the [`PHPCompatibilityWP`](https://github.com/PHPCompatibility/PHPCompatibilityWP) repository as a dependency and adds the associated PHPCS ruleset to the project ruleset.

PHPCompatibility(WP) always needs a `testVersion` to be set to know which issues should be flagged. The `testVersion` indicates the supported PHP versions of a project.

Based on input received on Slack, it has now been set to `7.2-` (= code has to be compatible with PHP 7.2 and higher).

Refs:
* https://github.com/PHPCompatibility/PHPCompatibilityWP
* https://github.com/PHPCompatibility/PHPCompatibility#sniffing-your-code-for-compatibility-with-specific-php-versions
* https://github.com/PHPCompatibility/PHPCompatibility#using-a-custom-ruleset