CS: partially fix embedded PHP
When PHP is embedded within HTML, single line snippets are fine.
However, for multi-line snippets of PHP, the PHP open and close tags should be on their own line, including when the snippet is the last thing in the file.

This commit addresses this for those multi-line embedded PHP snippets which did not start or end on the same line as an HTML snippet.

This partially fixes violations against the following sniffs:
* `Squiz.PHP.EmbeddedPhp`
* `WordPress.Arrays.ArrayIndentation`

This brings the number of violations down further to:
249 errors, 37 warnings (= 286) in 38 files with 15 different errorcodes

For the remaining violations, care should be taken that the HTML display is not affected by the changes which need to be made.
Pro-tip: quite often, the best solution will be moving multi-line PHP to before the HTML and assigning the result to a variable which is then echo'd out from within the HTML, but this will need to be evaluated on a case-by-case basis.