CS: fix inline control structures
Control structures should always use braces.
And then once braces are added, the content of the control structure should be on a new line, one indented and the closing brace should be on a line by itself.

Note: for one file, I've removed superfluous PHP open/close tags while fixing this as they are not needed, slow file processing and decrease code readability in this instance.

This fixes violations against the following sniffs:
* `Generic.ControlStructures.InlineControlStructures`

This brings the number of violations down further to:
205 errors, 37 warnings (= 242) in 34 files with 14 different errorcodes